+++ src/main/java/Project1/main.java
@@ -3,100 +3,6 @@ import java.io.*;
 import java.util.*;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-class Customer
-{
-    private String name;
-    private int point;
-    private boolean firsttime;
-    public Customer(String name)
-    {
-        this.name = name;
-        this.firsttime = true;
-        this.point = 0;
-    }
-    public void setName(String n){name =n;}
-    public void setPoints(int pt){point = pt;}//calculate from purchased order
-    public boolean getFirstTime(){return this.firsttime;}
-    public String getName(){return name;}
-    public int getPoints(){return point;}
-}
-class Product{
-    // Class variable
-    protected String product_code;
-    protected String name;
-    protected int unit_price;
-    // static HashMap to store total_sales of each products
-    int tol_sales_price;
-    int tol_sales_unit;
-
-    // Constructor
-    public Product(){ super();} // default
-    public Product(String pc, String name, int unit_price){
-        this.product_code = pc;
-        this.name = name;
-        this.unit_price = unit_price;
-        // If it is the first time that this product is mention create in the HashMap
-        tol_sales_price = 0;
-        tol_sales_unit = 0;
-    }
-
-    // Methods
-    // update the total_sales value
-    public void record_sales(int unit){
-        // Update the price by unit*unit_price
-        this.tol_sales_price += this.unit_price * unit;
-        this.tol_sales_unit += unit;
-    }
-
-    //Setter Getter
-    public int get_unit_price(){return this.unit_price;}
-    public String get_name(){return this.name;}
-    public String get_product_code(){return this.product_code;}
-
-    // Get total sales for product summary
-    public int get_tol_sales_price(){
-        return tol_sales_price;
-    }
-    public int get_tol_sales_unit() {
-        return tol_sales_unit;
-    }
-}
-class Installment
-{
-    private int months;
-    private double interest;
-  
-    public Installment(int month, double interest){this.months = month; this.interest = interest;}
-    public void setmonths(int m){months = m;}
-    public void setInterest(double i){interest = i;}
-    public int getMonths(){return this.months;}
-    public double getInterest(){return this.interest;}
-
-    public void print(){System.out.printf("%2d-month plan %4s monthly interest = %.2f%% \n",this.months," ",this.interest);}
-}
-class Order
-{
-    private int id;
-    private Customer customer;
-    private Product product;
-    private int unit;
-    private Installment installment;
-    public Order(){}
-    public Order(int i, Customer customer, Product product,int unit, Installment installment)
-    {
-        this.id = i;
-        this.customer = customer;
-        this.product = product;
-        this.unit = unit;
-        this.installment = installment;
-    }
-    public int getID(){return id;}
-    public Customer getCustomer(){return customer;}
-    public Product getProduct(){return product;}
-    public int getUnit(){return unit;}
-    public Installment getInstall(){return installment;}
-
-}
 
 class InvalidInputException extends Exception {
